backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
httpx==0.25.2
httpx[http2]==0.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
aiofiles==23.2.0
asyncio-throttle==1.0.2
pydantic[email]
bcrypt==3.2.2
aiohttp>=3.8.0